Update plu.sh
Added if statement to check if user is logged in as root; if so, give status message and exit.
Added another check to see if user was logged into the account used to create the node. If yes, then proceed; if no, then ask for the correct username. (This may be overkill, but don't want to take any chances)